// frontend/frontend/src/types.ts
export type IdParam = string | undefined | number;

export interface AcceptInvitationRequest {
    first_name: string;
    last_name: string;
    email: string;
    password: string;
    password_confirmation: string;
}

export interface RegisterAccountRequest extends AcceptInvitationRequest {
    locale: SupportedLocales;
}

export interface ResetPasswordRequest {
    password: string;
    password_confirmation: string;
}

export interface LoginResponse {
  access_token: string;       // Le token JWT d’accès
  refresh_token: string;      // Le token de refresh
  // si tu veux aussi envoyer user et accounts, il faut modifier ton backend pour les renvoyer
  user?: User;                // Optionnel, à ajouter si backend le renvoie
  accounts?: Account[];       // Optionnel, idem
}

export interface User {
    id?: IdParam;
    account_id?: IdParam;
    first_name: string;
    last_name: string;
    full_name: string;
    email: string;
    timezone?: string;
    password?: string;
    is_email_verified?: boolean;
    has_pending_email_change?: boolean;
    pending_email?: string;
    last_login_at?: string;
    status?: 'ACTIVE' | 'INACTIVE' | 'INVITED';
    role?: 'ADMIN' | 'ORGANIZER';
    is_account_owner?: boolean;
    locale?: SupportedLocales;
}

export interface Account {
    id?: IdParam;
    name: string;
    email: string;
    timezone?: string;
    currency_code?: string;
    password?: string;
    stripe_connect_setup_complete?: boolean;
    stripe_account_id?: string;
    is_account_email_confirmed?: boolean;
    is_saas_mode_enabled?: boolean;
    configuration?: AccountConfiguration;
    requires_manual_verification?: boolean;
}

export interface AccountConfiguration {
    id: IdParam;
    name: string;
    application_fees: {
        percentage: number;
        fixed: number;
    },
    is_system_default: boolean;
}
// ...
export interface GenericDataResponse<T> {
    data: T;
    errors?: Record<string, string>;
}